# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smartledger Premium API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin-users]
      summary: List all users
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResult"
    post:
      operationId: createUser
      tags: [admin-users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin-users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [admin-users]
      summary: Update user (balance, profit, status, etc)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [admin-users]
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [admin-users]
      summary: Suspend or unsuspend a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserInput"
      responses:
        "200":
          description: User status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # --- Deposits ---
...
  /admin/withdrawals:
    get:
      operationId: listWithdrawals
      tags: [admin-withdrawals]
      summary: List withdrawals
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalListResult"
  /admin/withdrawals/{id}:
    get:
      operationId: getWithdrawal
      tags: [admin-withdrawals]
      summary: Get withdrawal by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin-withdrawals]
      summary: Approve withdrawal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin-withdrawals]
      summary: Reject withdrawal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"

  # --- Transactions ---
  /admin/transactions:
    get:
      operationId: listTransactions
      tags: [admin-transactions]
      summary: List all transactions
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResult"

  # --- Investment Plans ---
...
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin-settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSettings
      tags: [admin-settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

  # --- Referrals ---
...
  /user/withdraw:
    post:
      operationId: submitWithdrawal
      tags: [user-actions]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, balance, profit, status, referralCode, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        balance: { type: number }
        profit: { type: number }
        status: { type: string }
        referralCode: { type: string }
        referredBy: { type: ["string", "null"] }
        createdAt: { type: string }
    UserListResult:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    UserInput:
      type: object
      required: [email, firstName, lastName, password]
      properties:
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        password: { type: string }
        phone: { type: string }
        country: { type: string }
        balance: { type: number }
        profit: { type: number }
    UserUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        country: { type: string }
        balance: { type: number }
        profit: { type: number }
        status: { type: string }
    SuspendUserInput:
      type: object
      required: [suspended]
      properties:
        suspended: { type: boolean }
        reason: { type: string }
...
    Withdrawal:
      type: object
      required: [id, userId, amount, method, walletAddress, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        amount: { type: number }
        method: { type: string }
        walletAddress: { type: string }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    WithdrawalListResult:
      type: object
      required: [withdrawals, total, page]
      properties:
        withdrawals:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total: { type: integer }
        page: { type: integer }
    WithdrawalInput:
      type: object
      required: [userId, amount, method, walletAddress]
      properties:
        userId: { type: integer }
        amount: { type: number }
        method: { type: string }
        walletAddress: { type: string }
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        type: { type: string }
        amount: { type: number }
        status: { type: string }
        description: { type: ["string", "null"] }
        createdAt: { type: string }
    TransactionListResult:
      type: object
      required: [transactions, total, page]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total: { type: integer }
        page: { type: integer }
...
    SiteSettings:
      type: object
      required: [siteName, adminEmail, referralBonusPercent, minDeposit, minWithdrawal]
      properties:
        siteName: { type: string }
        adminEmail: { type: string }
        referralBonusPercent: { type: number }
        minDeposit: { type: number }
        minWithdrawal: { type: number }
        maintenanceMode: { type: boolean }
        welcomeBonus: { type: number }
        smtpHost: { type: ["string", "null"] }
        smtpPort: { type: ["integer", "null"] }
        smtpUser: { type: ["string", "null"] }
    SiteSettingsUpdate:
      type: object
      properties:
        siteName: { type: string }
        adminEmail: { type: string }
        referralBonusPercent: { type: number }
        minDeposit: { type: number }
        minWithdrawal: { type: number }
        maintenanceMode: { type: boolean }
        welcomeBonus: { type: number }
        smtpHost: { type: string }
        smtpPort: { type: integer }
        smtpUser: { type: string }
        smtpPass: { type: string }
...
    ReviewNote:
      type: object
      properties:
        note: { type: string }
...