# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smartledger Premium API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  siteName: string;
  adminEmail: string;
  referralBonusPercent: number;
  minDeposit: number;
  minWithdrawal: number;
  maintenanceMode?: boolean;
  welcomeBonus?: number;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
}